DIFI Sink: Scaling args werent showing up and set min-max to default
The conditional was using the option_labels but it should be options.  I also think it will help to set the default to min-max mode, which has a default of -1 to +1, because the most common dtype in GR is complex, and typically values are between -1 and +1, so the average person trying out this DIFI block (e.g. with a Signal Source or a random SDR) is most likely going to be feeding the block values between -1 and +1, and if they don't read the docs carefully they are going to sit there and wonder why things aren't working, and it's not obvious what the problem is when you get 0's on the output.